get_date() function
Here I am creating a function to get the date. If the user is inputting data for today then all they need to do is enter Y and I am using the datetime module to get the correct date. If not then I use a input to get the day and month. split these into a list and convert them to int to create a new date.