insert_data_entry
Added the code which will be used to insert data entries into the SQL database. I need to collect all the variables in order to run the code which is my next task.

Also started the get_time function which will return the current time if the user is inputting the data right now. Need to add code to enter time manually if user is entering data from earlier in the day.