get_time() function
Working on a function to get the time the data is input into the database. If the user is inputting the data as it happens then it gathers the correct time from the datetime module. If the user is entering data from a previous time then it runs a series of code to get the correct time.