creating encrypted diary
connect_to_db() function takes 2 variables and connects to the database using a passphrase.

create_table() function creates a table if one doesn't exist. I noticed it does not update the table if i add an extra column.

main() is where I am keeping all of the code that will run. Need to create functions to collect the data and add them into the main function.

__name__ == __main__ the script will only run automatically if the python script is called directly. if the encrypted_dairy.py is imported in another script it will not be executed automatically