analysing data
creating a script to analyse the data in the ileostomy_freq database. started off by conecting to the databse using sqlite3 module. imported pandas as pd. created a dataframe by selecting all the data from the database. then checking for null values in the database and removing the rows if they contain null values. then using the describe() function to print some descriptive statistics of the database