histogram
created a function to plat a histogram. currently works for two columns I would plot a histogram for which is consistency and amount_percent. I want to look into other histograms such as using the dates as the x axis and how many entries on the y axis.

imported numpy and using arange to alter the x and y axis accordingly.