overlay histograms
working on a function to overlay multiple columns in a histogram. Currently working but the bars apear on top of each other. trying to figure out how to get the bars to appear side by side. GitHub CoPilot has suggested a 'bar' method. Online solutions don't seem as complicated but don't seem to work for me.