connection / get_date()
When i ran the script it was creating a new database in the parent folder rather than connecting to the original db. To solve this I used imported the os module and used the absolute path.

also the get_date() function when inputting data for another date was returning a different value than inputting data for todays date. to solve this I created a new variable and used .strftime to return the same value.